feat: 게시글 페이징  조회 및 날짜 범위 조회 기능 구현

+++ src/main/java/com/example/newsfeed/post/service/PostService.java
@@ -1,16 +1,20 @@
 package com.example.newsfeed.post.service;
 
-import com.example.newsfeed.entity.User;
+import com.example.newsfeed.post.dto.PostPageInfoResponseDto;
+import com.example.newsfeed.post.dto.PostPageResponseDto;
 import com.example.newsfeed.post.entity.Post;
 import com.example.newsfeed.post.dto.PostResponseDto;
 import com.example.newsfeed.post.repository.PostRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.sql.Timestamp;
 import java.util.List;
 
 @Service
@@ -21,47 +25,41 @@ public class PostService {
 //    private final UserRepository userRepository;
 
     // 게시글 생성
-    public PostResponseDto createPost(HttpServletRequest servletRequest, String title, String content) {
+    @Transactional
+    public PostResponseDto createPost(Long userId, String title, String content) {
 
-        // 기존 세션이 있으면 그 세션을 반환하고 없으면 null 을 반환
-        HttpSession session = servletRequest.getSession(false);
+//        User user = userRepository.findById(userId)
+//                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "유저가 존재하지 않습니다."));
 
-//        User loginUser = (User) session.getAttribute(Const.LOGIN_USER);
+//        Post post = new Post(title, content, user);
 
-//        if(loginUser == null){
-//            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "로그인한 사용자만 이용하실 수 있습니다.");
-//        }
+//        Post createPost = postRepository.save(post);
 
-        Post post = new Post(title, content);
-//        post.setUser(loginUser);
+//        return new PostResponseDto(
+//                createPost.getPostId(),
+//                createPost.getUser().getUsername(),
+//                createPost.getTitle(),
+//                createPost.getContent(),
+//                createPost.getLikesCount(),
+//                createPost.getCommentsCount(),
+//                createPost.getCreatedAt(),
+//                createPost.getModifiedAt());
 
-        Post createPost = postRepository.save(post);
-
-        return new PostResponseDto(
-                createPost.getPostId(),
-                createPost.getUser().getUsername(),
-                createPost.getTitle(),
-                createPost.getContent(),
-                createPost.getLikesCount(),
-                createPost.getCommentsCount(),
-                createPost.getCreatedAt(),
-                createPost.getModifiedAt());
+        return null; // <- 임시로 지정해놓은 리턴값, 위에 주석 해제 후 삭제
     }
 
     // 게시글 수정
-    public PostResponseDto updatePost(Long postId, HttpServletRequest servletRequest, String title, String content) {
+    @Transactional
+    public PostResponseDto updatePost(Long postId, Long userId, String title, String content) {
 
         // 입력받은 postId를 찾아서 post 변수에 저장
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "게시글을 찾을 수 없습니다."));
 
-        // 세션에서 이메일을 불러오고 업데이트할 스케줄을 쓴 유저와 이메일이 다르면 예외 발생
-        HttpSession session = servletRequest.getSession(false);
-//        User loginUser = (User) session.getAttribute(Const.LOGIN_USER);
-
-//        if(loginUser.getEmail().equals(post.getUser().getEmail())){
-//            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "로그인한 유저가 작성한 글이 없습니다.");
-//        }
+        // 지금 게시글을 쓴 유저 id 와 입력받은 유저 id 가 다르면 예외처리
+        if (!post.getUser().getUserId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "작성자만 게시글을 수정할 수 있습니다.");
+        }
 
         post.setTitle(title);
         post.setContent(content);
@@ -93,14 +91,43 @@ public class PostService {
                 .toList();
     }
 
-    // 게시글 페이징 조회(팔로우한 사람의 게시물을 최신순으로)
+    // 게시글 페이징 조회(게시물을 최신순으로) 추후에 팔로우한 사람들만 보이게 수정 예정
+    public PostPageInfoResponseDto<PostPageResponseDto> pagingPosts(int page, int size) {
+
+        Page<Post> postPage = postRepository.findAll(PageRequest.of(page - 1, size, Sort.by("createdAt").descending()));
+
+        List<PostPageResponseDto> list = postPage
+                .stream()
+                .map(PostPageResponseDto::toDto)
+                .toList();
+
+        return PostPageInfoResponseDto.toDto(
+                list,
+                page,
+                size
+        );
+    }
 
     // 게시글 날짜 범위로 검색
+    public List<PostResponseDto> searchPostsByDateRange(Timestamp startDate, Timestamp endDate) {
+
+        Sort sort = Sort.by(Sort.Direction.DESC, "modifiedAt");
+
+        return postRepository.findAllByCreatedAtBetween(startDate, endDate, sort)
+                .stream()
+                .map(PostResponseDto::toDto)
+                .toList();
+    }
 
     // 게시글 삭제
-    public void deletePost(Long postId) {
+    @Transactional
+    public void deletePost(Long postId, Long userId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "일정을 찾을 수 없습니다."));
+
+        if(!post.getUser().getUserId().equals(userId)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "작성자만 게시글을 삭제할 수 있습니다.");
+        }
 
         postRepository.delete(post);
     }
